[SQUASHED] Layers V2: Exposure of colors
Layers V2 is our take on Layers but this time, using CMTE exposure of hexes to make it easier for themers to port their themes and have much cleaner code than before.

This will also include a in-built Layers Manager into Settings but it will be simple and Layers Manager is still compatible. If some colors are still left to be exposed, it will be done in the next few weeks.

All credits to CM for this!
------------------------------------------------
Expose hardcoded colors

Opening up values for themers to have better control.

Change-Id: I3a0a460c5ffe0f5057b3b9ec92faa7a3e09c9e01

Conflicts:
	core/res/res/layout/resolver_list.xml

SystemUI: Expose values for themes.

Allow theming of volume panel separately from status. Allow
status bar icons to be tinted. Expose battery meter base color.
Expose status bar header elevation value. Add themeable ripple
color for navigation button press.

Change-Id: Ie3dbde82692a19927ba6c1da179e5d2f03391f6f

Conflicts:
	packages/SystemUI/res/values/cm_colors.xml
	packages/SystemUI/res/values/cm_dimens.xml

Themes: Allow notification shadows to be rounded

If a theme has a very large rounded rectangle it is obvious
the notification's shadow is a rectangle. So this patch
set the shadow to be drawn as a rounded rect like the notification itself.

Change-Id: I560fc5c4a4ce3d052d98b14af978a28c13b6caca

Conflicts:
	packages/SystemUI/res/values/cm_dimens.xml
	packages/SystemUI/src/com/android/systemui/statusbar/ExpandableOutlineView.java

SystemUI: Allow seperate theming of guts icons

Allows themes to assign different drawables for the notification
guts instead of using drawables used in other views.

Change-Id: I5f92e15de4aabde732e7cb89fd02432f32e8f329

Conflicts:
	packages/SystemUI/res/layout/notification_guts.xml